Move network related stuff into its own module
