+++ daemon/instructions.rs
@@ -5,7 +5,7 @@ use log::debug;
 
 use pueue::aliasing::insert_alias;
 use pueue::log::{clean_log_handles, read_and_compress_log_files};
-use pueue::message::*;
+use pueue::network::message::*;
 use pueue::state::{SharedState, State};
 use pueue::task::{Task, TaskStatus};
