+++ shared/lib.rs
@@ -1,10 +1,7 @@
 pub mod aliasing;
 pub mod log;
-pub mod message;
+pub mod network;
 pub mod platform;
-pub mod protocol;
-pub mod secret;
 pub mod settings;
 pub mod state;
 pub mod task;
-pub mod tls;
