+++ shared/network/secret.rs
